RDK-42997 WiFi Manager - Replace System Command - Fixing unit test issue
